feat(identityserver): implement core auth and registration features
- Added IdentityServer project and initial configuration
- Seeded data and created initial migration
- Implemented login and multi-step register page
- Configured client credentials flow for token requests
- Added custom profile service for claims handling
- Configured authentication and endpoints on the resource server

feat(gateway): setup gateway service with reverse proxy and auth

- Added Gateway service with reverse proxy configuration
- Integrated authentication into gateway config
- Tested gateway service with Postman
- Added remaining contracts and consumers to Auction and Search services
- Configured new client in Identity service for gateway integration

+++ src/SearchService/Entities/Item.cs
@@ -4,11 +4,11 @@ namespace SearchService.Entities
 {
     public class Item : Entity
     {
-        public int ReservePrice { get; set; }
+        public decimal ReservePrice { get; set; }
         public string Seller { get; set; }
         public string Winner { get; set; }
-        public int SoldAmount { get; set; }
-        public int CurrentHighBid { get; set; }
+        public decimal SoldAmount { get; set; }
+        public decimal CurrentHighBid { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
         public DateTime AuctionEnd { get; set; }
